Refactor Kanban board into modular components
